refactor: move inline tests into focused modules
